Added code for rendering & destroying resournces in  Irradiance Cube Map generation

+++ VKRenderer/VKRenderer/Core/Helper/Constants/Constans.hpp
@@ -202,8 +202,8 @@ struct FrameBufferDesc
 {
 	uint32_t							attachmentCount;
 	std::vector< VkImageView >			Attachments;
-	float								FBOWidth;
-	float								FBOHeight;
+	uint32_t							FBOWidth;
+	uint32_t							FBOHeight;
 	VkFramebuffer						FrameBuffer;
 };
